+++ src/command/create/artifacts/artifact-shared.ts
@@ -1,9 +1,8 @@
 /*
-* artifact-shared.ts
-*
-* Copyright (C) 2020-2022 Posit Software, PBC
-*
-*/
+ * artifact-shared.ts
+ *
+ * Copyright (C) 2020-2022 Posit Software, PBC
+ */
 
 import { capitalizeTitle } from "../../../core/text.ts";
 import { quartoConfig } from "../../../core/quarto.ts";
@@ -16,24 +15,7 @@ import { basename, dirname, join, relative } from "path/mod.ts";
 import { ensureDirSync, walkSync } from "fs/mod.ts";
 import { renderEjs } from "../../../core/ejs.ts";
 import { safeExistsSync } from "../../../core/path.ts";
-
-export interface CreateDirective {
-  displayType: string;
-  name: string;
-  directory: string;
-  template: string;
-  options: Record<string, unknown>;
-}
-
-export interface CreateDirectiveData extends Record<string, string> {
-  name: string;
-  filesafename: string;
-  classname: string;
-  title: string;
-  author: string;
-  version: string;
-  quartoversion: string;
-}
+import { CreateDirective, CreateDirectiveData } from "../cmd-types.ts";
 
 // File paths that include this string will get fixed up
 // and the value from the ejs data will be substituted
